# Add RB_UNLIKELY hint in dump method for tracing log