--- a/ext/oj/dump_compat.c
+++ b/ext/oj/dump_compat.c
@@ -109,7 +109,7 @@ dump_to_json(VALUE obj, Out out) {
     const char		*s;
     int			len;
 
-    if (Yes == out->opts->trace) {
+    if (RB_UNLIKELY(Yes == out->opts->trace)) {
 	oj_trace("to_json", obj, __FILE__, __LINE__, 0, TraceRubyIn);
     }
     if (0 == rb_obj_method_arity(obj, oj_to_json_id)) {
@@ -117,7 +117,7 @@ dump_to_json(VALUE obj, Out out) {
     } else {
 	rs = rb_funcall2(obj, oj_to_json_id, out->argc, out->argv);
     }
-    if (Yes == out->opts->trace) {
+    if (RB_UNLIKELY(Yes == out->opts->trace)) {
 	oj_trace("to_json", obj, __FILE__, __LINE__, 0, TraceRubyOut);
     }
 
@@ -893,7 +893,7 @@ void
 oj_dump_compat_val(VALUE obj, int depth, Out out, bool as_ok) {
     int	type = rb_type(obj);
 
-    if (Yes == out->opts->trace) {
+    if (RB_UNLIKELY(Yes == out->opts->trace)) {
 	oj_trace("dump", obj, __FILE__, __LINE__, depth, TraceIn);
     }
     if (out->opts->dump_opts.max_depth <= depth) {
@@ -918,14 +918,14 @@ oj_dump_compat_val(VALUE obj, int depth, Out out, bool as_ok) {
 
 	if (NULL != f) {
 	    f(obj, depth, out, as_ok);
-	    if (Yes == out->opts->trace) {
+	    if (RB_UNLIKELY(Yes == out->opts->trace)) {
 		oj_trace("dump", obj, __FILE__, __LINE__, depth, TraceOut);
 	    }
 	    return;
 	}
     }
     oj_dump_nil(Qnil, depth, out, false);
-    if (Yes == out->opts->trace) {
+    if (RB_UNLIKELY(Yes == out->opts->trace)) {
 	oj_trace("dump", Qnil, __FILE__, __LINE__, depth, TraceOut);
     }
 }