If the compiler supports it, the UNLIKELY hint might speed up the conditional branching slightly.
Ref. https://github.com/ohler55/oj/pull/782

−       | before  | after   | result
--       | --      | --      | --
Oj.dump  | 11.754M | 11.775M | -

### Environment
- Linux
  - Manjaro Linux x86_64
  - Kernel: 5.18.12-3-MANJARO
  - AMD Ryzen 7 5700G
  - gcc version 12.1.0
  - Ruby 3.1.2

### Before
```
Warming up --------------------------------------
             Oj.dump     1.166M i/100ms
Calculating -------------------------------------
             Oj.dump     11.754M (± 0.2%) i/s -    177.254M in  15.080998s
```

### After
```
Warming up --------------------------------------
             Oj.dump     1.184M i/100ms
Calculating -------------------------------------
             Oj.dump     11.775M (± 0.2%) i/s -    177.612M in  15.084141s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

data = 42

Benchmark.ips do |x|
  x.warmup = 5
  x.time = 15

  x.report('Oj.dump') do |times|
    i = 0
    while i < times
      Oj.dump(data)
      i += 1
    end
  end
end
```